Shader updates
-Made a separate file for shaders
-added some stage related shaders
